Move fce.php to /libs/
