+++ www/index.php
@@ -1,6 +1,6 @@
 <?php
 declare(strict_types=1);
-require 'fce.php';
+require 'libs/fce.php';
 
 $items = [
     0 => "Lorem ipsum dolor sit amet, consectetuer adipiscing elit.",
